+++ tests/testthat/test_Unary.R
@@ -316,6 +316,94 @@ test_that("identize_byname works as expected", {
 
 
 ###########################################################
+context("Vectorize")
+###########################################################
+
+test_that("vectorize_byname works as expected", {
+  # Try with a square matrix
+  m1 <- matrix(c(1, 5,
+                4, 5),
+              nrow = 2, ncol = 2, byrow = TRUE, 
+              dimnames = list(c("p1", "p2"), c("i1", "i2"))) %>% 
+    setrowtype("Products") %>% setcoltype("Industries")
+  expected1 <- matrix(c(1, 
+                        4, 
+                        5, 
+                        5),
+                      nrow = 4, ncol = 1, 
+                      dimnames = list(c("p1 -> i1", "p2 -> i1", "p1 -> i2", "p2 -> i2"))) %>% 
+    setrowtype("Products") %>% setcoltype("Industries")
+  actual1 <- vectorize_byname(m1)
+  expect_equal(actual1, expected1)
+  # Try with a rectangular matrix
+  m2 <- matrix(c(1, 2, 3,
+                 4, 5, 6),
+               nrow = 2, ncol = 3, byrow = TRUE,
+               dimnames = list(c("p1", "p2"), c("i1", "i2", "i3"))) %>% 
+    setrowtype("Products") %>% setcoltype("Industries")
+  expected2 <- matrix(c(1, 
+                       4, 
+                       2, 
+                       5, 
+                       3, 
+                       6),
+                     nrow = 6, ncol = 1,
+                     dimnames = list(c("p1 -> i1", "p2 -> i1", "p1 -> i2", "p2 -> i2", "p1 -> i3", "p2 -> i3"))) %>% 
+    setrowtype("Products") %>% setcoltype("Industries")
+  actual2 <- vectorize_byname(m2)
+  expect_equal(actual2, expected2)
+  # Try with a single number
+  m3 <- 42
+  expected3 <- m3
+  dim(expected3) <- c(1, 1)
+  dimnames(expected3) <- list(c(NULL))
+  actual3 <- vectorize_byname(m3)
+  expect_equal(actual3, expected3)
+  # Try with a different separator
+  m4 <- matrix(c(1, 5,
+                 4, 5),
+               nrow = 2, ncol = 2, byrow = TRUE, 
+               dimnames = list(c("p1", "p2"), c("i1", "i2"))) %>% 
+    setrowtype("Products") %>% setcoltype("Industries")
+  expected4 <- matrix(c(1, 
+                        4, 
+                        5, 
+                        5),
+                      nrow = 4, ncol = 1, 
+                      dimnames = list(c("p1---i1", "p2---i1", "p1---i2", "p2---i2"))) %>% 
+    setrowtype("Products") %>% setcoltype("Industries")
+  actual4 <- vectorize_byname(m4, sep = "---")
+  expect_equal(actual4, expected4)
+  # Test with a matrix that is already a column vector
+  m5 <- matrix(c(1,
+                 2,
+                 3),
+               nrow = 3, ncol = 1,
+               dimnames = list(c("p1", "p2", "p3"), "i1")) %>% 
+    setrowtype("Products") %>% setcoltype("Industries")
+  actual5 <- vectorize_byname(m5, sep = "***")
+  expected5 <- matrix(c(1,
+                        2,
+                        3),
+                      nrow = 3, ncol = 1,
+                      dimnames = list(c("p1***i1", "p2***i1", "p3***i1"))) %>% 
+    setrowtype("Products") %>% setcoltype("Industries")
+  expect_equal(actual5, expected5)
+  # Test with NULL. Should get NULL back.
+  expect_true(is.null(vectorize_byname(NULL)))
+  # Test with NA.
+  expect_error(vectorize_byname(NA), "a is not numeric in vectorize_byname")
+  # Test with string
+  expect_error(vectorize_byname("a"), "a is not numeric in vectorize_byname")
+  # Test with a list of matrices
+  list6 <- list(m1, m1)
+  actual6 <- vectorize_byname(list6)
+  expected6 <- list(expected1, expected1)
+  expect_equal(actual6, expected6)
+})
+
+
+###########################################################
 context("Fractionize")
 ###########################################################