+++ vignettes/matsbyname.R
@@ -28,7 +28,7 @@ U + Y2
 ## -----------------------------------------------------------------------------
 Y3 <- matrix(5:8, ncol = 2, dimnames = list(c("p1", "p3"), c("i1", "i3")))
 Y3
-# Nonsensical because neither row nor column names are respected. 
+# Nonsensical because neither row names nor column names are respected. 
 # The "p3" rows and "i3" columns of Y3 have been added to 
 # "p2" rows and "i2" columns of U.
 # Row and column names for the sum are taken from the first operand (U).
@@ -105,4 +105,49 @@ tryCatch(matrixproduct_byname(A, B), error = function(err){print(err)})
 sum_byname(A, list(B, B))
 hadamardproduct_byname(list(A, A), B)
 matrixproduct_byname(list(A, A), list(C, C))
+
+## -----------------------------------------------------------------------------
+tidy <- data.frame(
+  matrix = c("A", "A", "A", "A", "B", "B", "B", "B"),
+  row = c("p1", "p1", "p2", "p2", "p1", "p1", "p2", "p2"),
+  col = c("i1", "i2", "i1", "i2", "i1", "i2", "i1", "i2"),
+  vals = c(1, 3, 2, 4, 8, 6, 7, 5)
+) %>%
+  mutate(
+    rowtype = "Industries",
+    coltype  = "Products"
+  )
+tidy
+mats <- tidy %>%
+  group_by(matrix) %>%
+  matsindf::collapse_to_matrices(matnames = "matrix", matvals = "vals",
+                                 rownames = "row", colnames = "col",
+                                 rowtypes = "rowtype", coltypes = "coltype") %>%
+  rename(
+    matrix.name = matrix,
+    matrix = vals
+  )
+mats
+mats$matrix[[1]]
+mats$matrix[[2]]
+
+## -----------------------------------------------------------------------------
+result <- mats %>%
+  tidyr::spread(key = matrix.name, value = matrix) %>%
+  # Duplicate the row to demonstrate byname operating simultaneously
+  # on all rows of the data frame.
+  dplyr::bind_rows(., .) %>%
+  dplyr::mutate(
+    # Create a column of constants.
+    c = make_list(x = 1:2, n = 2, lenx = 2),
+    # Sum all rows of the data frame with a single instruction.
+    sum = sum_byname(A, B),
+    # Multiply matrices in the sum column by corresponding constants in the c column.
+    product = hadamardproduct_byname(c, sum)
+  )
+result
+result$sum[[1]]
+result$sum[[2]]
+result$product[[1]]
+result$product[[2]]